Update README.md
"I am a developer of web and desktop applications, as well as 2D games, specializing in Python. I studied at Tokio School, an institution headquartered in Lisbon with a branch in Porto, Portugal. As a freelancer, I provide solutions for small and medium-sized businesses, as well as independent professionals, focusing on task automation and asset and personnel management.

My goal for 2025 is to continue enhancing my solutions to improve employee and management productivity, delivering greater efficiency and practicality in business processes."